pyproject.toml
[tool.poetry]
name = "sber-test"
version = "0.1.0"
description = "Тестовое задание для Сбера"
authors = ["komediantto <komediantto@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2"
djangorestframework = "^3.14.0"
python-dateutil = "^2.8.2"
coverage = "^7.2.3"
python-dotenv = "^1.0.0"


[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
